enhance: improve naming and Icon selecting logic for tools
